refactor(analysis): extract call traversal and reference counting from server.rs
Move two large AST-walking families out of hew-lsp/src/server.rs and
into hew-analysis where they belong:

- hew-analysis/src/calls.rs (new): CallSite struct and
  collect_calls_in_block/stmt/expr. Used by the LSP for call hierarchy
  (incoming/outgoing) and code lens. ~260 lines moved from server.rs.

- hew-analysis/src/references.rs: count_all_references and the private
  count_idents_in_item/block/stmt/expr helpers. Used by the LSP for
  code-lens reference-count hints. ~260 lines moved from server.rs.

server.rs now imports these via hew_analysis::calls::collect_calls_in_block
and hew_analysis::references::count_all_references. Five AST types
(Block, Expr, Stmt, StringPart, TraitItem, TypeBodyItem) are no longer
imported directly since they were only needed by the removed code.

server.rs shrinks by ~520 lines and becomes closer to pure protocol
handling. Both extracted modules are available to hew-wasm if call
hierarchy or reference counting is needed there in future.
