+++ hew-analysis/src/lib.rs
@@ -4,6 +4,7 @@
 //! `hew-lsp` (native LSP server) and `hew-wasm` (browser-based editor support).
 //! All result types use plain offsets and strings rather than LSP protocol types.
 
+pub mod calls;
 pub mod code_actions;
 pub mod completions;
 pub mod definition;
