+++ hew-lsp/src/server.rs
@@ -3,10 +3,9 @@
 use std::collections::{HashMap, HashSet};
 
 use dashmap::DashMap;
-use hew_parser::ast::{
-    Attribute, Block, Expr, ImportDecl, ImportSpec, Item, Span, Stmt, StringPart, TraitItem,
-    TypeBodyItem, TypeDeclKind,
-};
+use hew_analysis::calls::collect_calls_in_block;
+use hew_analysis::references::count_all_references;
+use hew_parser::ast::{Attribute, ImportDecl, ImportSpec, Item, Span, TypeDeclKind};
 use hew_parser::ParseResult;
 use hew_types::error::{Severity, TypeErrorKind};
 use hew_types::module_registry::build_module_search_paths;
@@ -2710,272 +2709,6 @@ fn find_callable_at(
     None
 }
 
-struct CallSite {
-    name: String,
-    span: Span,
-}
-
-fn collect_calls_in_block(block: &Block, calls: &mut Vec<CallSite>) {
-    for (stmt, _span) in &block.stmts {
-        collect_calls_in_stmt(stmt, calls);
-    }
-}
-
-fn collect_calls_in_stmt(stmt: &Stmt, calls: &mut Vec<CallSite>) {
-    match stmt {
-        Stmt::Let { value, .. } | Stmt::Var { value, .. } => {
-            if let Some(spanned) = value {
-                collect_calls_in_expr(spanned, calls);
-            }
-        }
-        Stmt::Expression(spanned) | Stmt::Return(Some(spanned)) => {
-            collect_calls_in_expr(spanned, calls);
-        }
-        Stmt::Defer(d) => {
-            collect_calls_in_expr(d.as_ref(), calls);
-        }
-        Stmt::Assign { target, value, .. } => {
-            collect_calls_in_expr(target, calls);
-            collect_calls_in_expr(value, calls);
-        }
-        Stmt::For { iterable, body, .. } => {
-            collect_calls_in_expr(iterable, calls);
-            collect_calls_in_block(body, calls);
-        }
-        Stmt::While {
-            condition, body, ..
-        } => {
-            collect_calls_in_expr(condition, calls);
-            collect_calls_in_block(body, calls);
-        }
-        Stmt::WhileLet { expr, body, .. } => {
-            collect_calls_in_expr(expr.as_ref(), calls);
-            collect_calls_in_block(body, calls);
-        }
-        Stmt::If {
-            condition,
-            then_block,
-            else_block,
-            ..
-        } => {
-            collect_calls_in_expr(condition, calls);
-            collect_calls_in_block(then_block, calls);
-            if let Some(eb) = else_block {
-                if let Some(if_box) = &eb.if_stmt {
-                    let (if_stmt, _) = if_box.as_ref();
-                    collect_calls_in_stmt(if_stmt, calls);
-                }
-                if let Some(b) = &eb.block {
-                    collect_calls_in_block(b, calls);
-                }
-            }
-        }
-        Stmt::IfLet {
-            expr,
-            body,
-            else_body,
-            ..
-        } => {
-            collect_calls_in_expr(expr.as_ref(), calls);
-            collect_calls_in_block(body, calls);
-            if let Some(block) = else_body {
-                collect_calls_in_block(block, calls);
-            }
-        }
-        Stmt::Match { scrutinee, arms } => {
-            collect_calls_in_expr(scrutinee, calls);
-            for arm in arms {
-                if let Some(g) = &arm.guard {
-                    collect_calls_in_expr(g, calls);
-                }
-                collect_calls_in_expr(&arm.body, calls);
-            }
-        }
-        Stmt::Loop { body, .. } => collect_calls_in_block(body, calls),
-        Stmt::Return(None) | Stmt::Break { .. } | Stmt::Continue { .. } => {}
-    }
-}
-
-#[expect(
-    clippy::too_many_lines,
-    reason = "exhaustive match over all Expr variants"
-)]
-fn collect_calls_in_expr(spanned: &(Expr, Span), calls: &mut Vec<CallSite>) {
-    let (expr, expr_span) = spanned;
-    match expr {
-        Expr::Call { function, args, .. } => {
-            let (func, _) = function.as_ref();
-            if let Expr::Identifier(name) = func {
-                calls.push(CallSite {
-                    name: name.clone(),
-                    span: expr_span.clone(),
-                });
-            }
-            collect_calls_in_expr(function.as_ref(), calls);
-            for arg in args {
-                collect_calls_in_expr(arg.expr(), calls);
-            }
-        }
-        Expr::MethodCall {
-            receiver,
-            method,
-            args,
-            ..
-        } => {
-            calls.push(CallSite {
-                name: method.clone(),
-                span: expr_span.clone(),
-            });
-            collect_calls_in_expr(receiver.as_ref(), calls);
-            for arg in args {
-                collect_calls_in_expr(arg.expr(), calls);
-            }
-        }
-        Expr::IfLet {
-            expr,
-            body,
-            else_body,
-            ..
-        } => {
-            collect_calls_in_expr(expr.as_ref(), calls);
-            collect_calls_in_block(body, calls);
-            if let Some(block) = else_body {
-                collect_calls_in_block(block, calls);
-            }
-        }
-        Expr::Send { target, message } => {
-            // Track actor message sends as call edges
-            let send_name = match &target.as_ref().0 {
-                Expr::Identifier(name) => format!("{name}.send"),
-                _ => "send".to_string(),
-            };
-            calls.push(CallSite {
-                name: send_name,
-                span: expr_span.clone(),
-            });
-            collect_calls_in_expr(target.as_ref(), calls);
-            collect_calls_in_expr(message.as_ref(), calls);
-        }
-        Expr::Binary { left, right, .. } => {
-            collect_calls_in_expr(left.as_ref(), calls);
-            collect_calls_in_expr(right.as_ref(), calls);
-        }
-        Expr::Unary { operand, .. } => {
-            collect_calls_in_expr(operand.as_ref(), calls);
-        }
-        Expr::Await(a) => {
-            collect_calls_in_expr(a.as_ref(), calls);
-        }
-        Expr::PostfixTry(p) => {
-            collect_calls_in_expr(p.as_ref(), calls);
-        }
-        Expr::Cast { expr, .. } => {
-            collect_calls_in_expr(expr.as_ref(), calls);
-        }
-        Expr::Yield(Some(y)) => {
-            collect_calls_in_expr(y.as_ref(), calls);
-        }
-        Expr::If {
-            condition,
-            then_block,
-            else_block,
-            ..
-        } => {
-            collect_calls_in_expr(condition.as_ref(), calls);
-            collect_calls_in_expr(then_block.as_ref(), calls);
-            if let Some(eb) = else_block {
-                collect_calls_in_expr(eb.as_ref(), calls);
-            }
-        }
-        Expr::Match { scrutinee, arms } => {
-            collect_calls_in_expr(scrutinee.as_ref(), calls);
-            for arm in arms {
-                if let Some(g) = &arm.guard {
-                    collect_calls_in_expr(g, calls);
-                }
-                collect_calls_in_expr(&arm.body, calls);
-            }
-        }
-        Expr::Block(b) | Expr::Unsafe(b) => collect_calls_in_block(b, calls),
-        Expr::Index { object, index, .. } => {
-            collect_calls_in_expr(object.as_ref(), calls);
-            collect_calls_in_expr(index.as_ref(), calls);
-        }
-        Expr::FieldAccess { object, .. } => {
-            collect_calls_in_expr(object.as_ref(), calls);
-        }
-        Expr::Lambda { body, .. } | Expr::SpawnLambdaActor { body, .. } => {
-            collect_calls_in_expr(body.as_ref(), calls);
-        }
-        Expr::ArrayRepeat { value, count } => {
-            collect_calls_in_expr(value.as_ref(), calls);
-            collect_calls_in_expr(count.as_ref(), calls);
-        }
-        Expr::MapLiteral { entries } => {
-            for (k, v) in entries {
-                collect_calls_in_expr(k, calls);
-                collect_calls_in_expr(v, calls);
-            }
-        }
-        Expr::Tuple(exprs) | Expr::Array(exprs) | Expr::Join(exprs) => {
-            for e in exprs {
-                collect_calls_in_expr(e, calls);
-            }
-        }
-        Expr::Range { start, end, .. } => {
-            if let Some(s) = start {
-                collect_calls_in_expr(s.as_ref(), calls);
-            }
-            if let Some(e) = end {
-                collect_calls_in_expr(e.as_ref(), calls);
-            }
-        }
-        Expr::InterpolatedString(parts) => {
-            for part in parts {
-                if let StringPart::Expr(e) = part {
-                    collect_calls_in_expr(e, calls);
-                }
-            }
-        }
-        Expr::Scope { body, .. } | Expr::ScopeLaunch(body) | Expr::ScopeSpawn(body) => {
-            collect_calls_in_block(body, calls);
-        }
-        Expr::Select { arms, timeout, .. } => {
-            for arm in arms {
-                collect_calls_in_expr(&arm.body, calls);
-            }
-            if let Some(tc) = timeout {
-                collect_calls_in_expr(tc.duration.as_ref(), calls);
-                collect_calls_in_expr(tc.body.as_ref(), calls);
-            }
-        }
-        Expr::StructInit { fields, .. } => {
-            for (_, v) in fields {
-                collect_calls_in_expr(v, calls);
-            }
-        }
-        Expr::Spawn { target, args } => {
-            collect_calls_in_expr(target.as_ref(), calls);
-            for (_, a) in args {
-                collect_calls_in_expr(a, calls);
-            }
-        }
-        Expr::Timeout { expr, duration } => {
-            collect_calls_in_expr(expr.as_ref(), calls);
-            collect_calls_in_expr(duration.as_ref(), calls);
-        }
-        Expr::Literal(_)
-        | Expr::Identifier(_)
-        | Expr::Cooperate
-        | Expr::ScopeCancel
-        | Expr::This
-        | Expr::RegexLiteral(_)
-        | Expr::ByteStringLiteral(_)
-        | Expr::ByteArrayLiteral(_)
-        | Expr::Yield(None) => {}
-    }
-}
-
 fn find_incoming_calls(
     uri: &Url,
     source: &str,
@@ -3149,271 +2882,6 @@ fn build_code_lenses(source: &str, lo: &[usize], parse_result: &ParseResult) -> 
 
 fn has_test_attribute(attrs: &[Attribute]) -> bool {
     attrs.iter().any(|a| a.name == "test")
-}
-
-fn count_all_references(parse_result: &ParseResult) -> HashMap<String, usize> {
-    let mut counts = HashMap::new();
-    for (item, _) in &parse_result.program.items {
-        count_idents_in_item(item, &mut counts);
-    }
-    counts
-}
-
-fn count_idents_in_item(item: &Item, counts: &mut HashMap<String, usize>) {
-    match item {
-        Item::Function(f) => count_idents_in_block(&f.body, counts),
-        Item::Actor(a) => {
-            if let Some(init) = &a.init {
-                count_idents_in_block(&init.body, counts);
-            }
-            if let Some(term) = &a.terminate {
-                count_idents_in_block(&term.body, counts);
-            }
-            for recv in &a.receive_fns {
-                count_idents_in_block(&recv.body, counts);
-            }
-            for method in &a.methods {
-                count_idents_in_block(&method.body, counts);
-            }
-        }
-        Item::TypeDecl(td) => {
-            for body_item in &td.body {
-                if let TypeBodyItem::Method(m) = body_item {
-                    count_idents_in_block(&m.body, counts);
-                }
-            }
-        }
-        Item::Impl(i) => {
-            for method in &i.methods {
-                count_idents_in_block(&method.body, counts);
-            }
-        }
-        Item::Trait(t) => {
-            for trait_item in &t.items {
-                if let TraitItem::Method(m) = trait_item {
-                    if let Some(body) = &m.body {
-                        count_idents_in_block(body, counts);
-                    }
-                }
-            }
-        }
-        Item::Const(c) => count_idents_in_expr(&c.value.0, counts),
-        Item::Import(_)
-        | Item::ExternBlock(_)
-        | Item::Wire(_)
-        | Item::TypeAlias(_)
-        | Item::Supervisor(_)
-        | Item::Machine(_) => {}
-    }
-}
-
-fn count_idents_in_block(block: &Block, counts: &mut HashMap<String, usize>) {
-    for (stmt, _) in &block.stmts {
-        count_idents_in_stmt(stmt, counts);
-    }
-    if let Some(trailing) = &block.trailing_expr {
-        count_idents_in_expr(&trailing.0, counts);
-    }
-}
-
-fn count_idents_in_stmt(stmt: &Stmt, counts: &mut HashMap<String, usize>) {
-    match stmt {
-        Stmt::Let { value, .. } | Stmt::Var { value, .. } => {
-            if let Some(val) = value {
-                count_idents_in_expr(&val.0, counts);
-            }
-        }
-        Stmt::Assign { target, value, .. } => {
-            count_idents_in_expr(&target.0, counts);
-            count_idents_in_expr(&value.0, counts);
-        }
-        Stmt::If {
-            condition,
-            then_block,
-            else_block,
-        } => {
-            count_idents_in_expr(&condition.0, counts);
-            count_idents_in_block(then_block, counts);
-            if let Some(eb) = else_block {
-                if let Some(if_stmt) = &eb.if_stmt {
-                    count_idents_in_stmt(&if_stmt.0, counts);
-                }
-                if let Some(block) = &eb.block {
-                    count_idents_in_block(block, counts);
-                }
-            }
-        }
-        Stmt::IfLet {
-            expr,
-            body,
-            else_body,
-            ..
-        } => {
-            count_idents_in_expr(&expr.0, counts);
-            count_idents_in_block(body, counts);
-            if let Some(block) = else_body {
-                count_idents_in_block(block, counts);
-            }
-        }
-        Stmt::Match { scrutinee, arms } => {
-            count_idents_in_expr(&scrutinee.0, counts);
-            for arm in arms {
-                if let Some(guard) = &arm.guard {
-                    count_idents_in_expr(&guard.0, counts);
-                }
-                count_idents_in_expr(&arm.body.0, counts);
-            }
-        }
-        Stmt::Loop { body, .. }
-        | Stmt::While { body, .. }
-        | Stmt::WhileLet { body, .. }
-        | Stmt::For { body, .. } => {
-            count_idents_in_block(body, counts);
-        }
-        Stmt::Expression(expr) | Stmt::Return(Some(expr)) => {
-            count_idents_in_expr(&expr.0, counts);
-        }
-        Stmt::Break {
-            value: Some(val), ..
-        } => count_idents_in_expr(&val.0, counts),
-        _ => {}
-    }
-}
-
-#[expect(
-    clippy::too_many_lines,
-    reason = "exhaustive match over all Expr variants for identifier counting"
-)]
-fn count_idents_in_expr(expr: &Expr, counts: &mut HashMap<String, usize>) {
-    match expr {
-        Expr::Identifier(ident) => {
-            *counts.entry(ident.clone()).or_insert(0) += 1;
-        }
-        Expr::Binary { left, right, .. } => {
-            count_idents_in_expr(&left.0, counts);
-            count_idents_in_expr(&right.0, counts);
-        }
-        Expr::Unary { operand, .. } => count_idents_in_expr(&operand.0, counts),
-        Expr::Call { function, args, .. } => {
-            count_idents_in_expr(&function.0, counts);
-            for arg in args {
-                count_idents_in_expr(&arg.expr().0, counts);
-            }
-        }
-        Expr::MethodCall { receiver, args, .. } => {
-            count_idents_in_expr(&receiver.0, counts);
-            for arg in args {
-                count_idents_in_expr(&arg.expr().0, counts);
-            }
-        }
-        Expr::FieldAccess { object, .. } => count_idents_in_expr(&object.0, counts),
-        Expr::Index { object, index } => {
-            count_idents_in_expr(&object.0, counts);
-            count_idents_in_expr(&index.0, counts);
-        }
-        Expr::StructInit { fields, .. } => {
-            for (_, val) in fields {
-                count_idents_in_expr(&val.0, counts);
-            }
-        }
-        Expr::Spawn { target, args } => {
-            count_idents_in_expr(&target.0, counts);
-            for (_, val) in args {
-                count_idents_in_expr(&val.0, counts);
-            }
-        }
-        Expr::Block(block)
-        | Expr::Unsafe(block)
-        | Expr::ScopeLaunch(block)
-        | Expr::ScopeSpawn(block) => count_idents_in_block(block, counts),
-        Expr::Scope { body, .. } => count_idents_in_block(body, counts),
-        Expr::If {
-            condition,
-            then_block,
-            else_block,
-        } => {
-            count_idents_in_expr(&condition.0, counts);
-            count_idents_in_expr(&then_block.0, counts);
-            if let Some(else_expr) = else_block {
-                count_idents_in_expr(&else_expr.0, counts);
-            }
-        }
-        Expr::IfLet {
-            expr,
-            body,
-            else_body,
-            ..
-        } => {
-            count_idents_in_expr(&expr.0, counts);
-            count_idents_in_block(body, counts);
-            if let Some(block) = else_body {
-                count_idents_in_block(block, counts);
-            }
-        }
-        Expr::Match { scrutinee, arms } => {
-            count_idents_in_expr(&scrutinee.0, counts);
-            for arm in arms {
-                if let Some(guard) = &arm.guard {
-                    count_idents_in_expr(&guard.0, counts);
-                }
-                count_idents_in_expr(&arm.body.0, counts);
-            }
-        }
-        Expr::Lambda { body, .. } | Expr::SpawnLambdaActor { body, .. } => {
-            count_idents_in_expr(&body.0, counts);
-        }
-        Expr::ArrayRepeat { value, count } => {
-            count_idents_in_expr(&value.0, counts);
-            count_idents_in_expr(&count.0, counts);
-        }
-        Expr::MapLiteral { entries } => {
-            for (k, v) in entries {
-                count_idents_in_expr(&k.0, counts);
-                count_idents_in_expr(&v.0, counts);
-            }
-        }
-        Expr::Tuple(elems) | Expr::Array(elems) | Expr::Join(elems) => {
-            for elem in elems {
-                count_idents_in_expr(&elem.0, counts);
-            }
-        }
-        Expr::Send { target, message } => {
-            count_idents_in_expr(&target.0, counts);
-            count_idents_in_expr(&message.0, counts);
-        }
-        Expr::Select {
-            arms: sel_arms,
-            timeout,
-        } => {
-            for arm in sel_arms {
-                count_idents_in_expr(&arm.body.0, counts);
-            }
-            if let Some(t) = timeout {
-                count_idents_in_expr(&t.body.0, counts);
-            }
-        }
-        Expr::Timeout { expr: e, .. } => count_idents_in_expr(&e.0, counts),
-        Expr::Await(inner) | Expr::PostfixTry(inner) | Expr::Yield(Some(inner)) => {
-            count_idents_in_expr(&inner.0, counts);
-        }
-        Expr::Cast { expr: inner, .. } => count_idents_in_expr(&inner.0, counts),
-        Expr::Range { start, end, .. } => {
-            if let Some(s) = start {
-                count_idents_in_expr(&s.0, counts);
-            }
-            if let Some(e) = end {
-                count_idents_in_expr(&e.0, counts);
-            }
-        }
-        Expr::InterpolatedString(parts) => {
-            for part in parts {
-                if let StringPart::Expr(e) = part {
-                    count_idents_in_expr(&e.0, counts);
-                }
-            }
-        }
-        _ => {}
-    }
 }
 
 // ── Workspace symbol helpers ────────────────────────────────────────
